requirements.txt
# Web framework & server
fastapi==0.136.0
uvicorn[standard]==0.46.0
python-multipart==0.0.26   # file upload support for FastAPI

# Config
python-dotenv==1.2.2

# Database
aiosqlite==0.22.1           # async SQLite3 access

# Auth (Google OAuth 2.0)
authlib==1.7.0
httpx==0.28.1               # HTTP client (OAuth flows + Ollama API calls)
itsdangerous==2.2.0         # session signing

# PDF parsing
pymupdf==1.27.2.2           # fitz: extract clue text from PDF
pdf2image==1.17.0           # convert PDF page to image (requires poppler system pkg)
opencv-python==4.13.0.92    # grid detection via Hough lines + contours
pillow==12.2.0              # image handling
numpy==2.4.4                # array operations for grid classification

# previously listed, no longer the primary parser libs but kept for reference
pdfplumber==0.11.9
pypdf==6.10.2
